+++ Receptor/Client/src/modelo_r/InternetManager.java
@@ -1,4 +1,4 @@
-package modelo;
+package modelo_r;
 
 import java.io.BufferedReader;
 import java.io.IOException;
