+++ Receptor/Client/src/modelo_r/alarma/GestorArchivoAudio.java
@@ -1,4 +1,4 @@
-package modelo.alarma;
+package modelo_r.alarma;
 
 import java.io.File;
 import java.io.IOException;
