+++ Receptor/Client/src/modelo_r/alarma/HiloDeAudio.java
@@ -1,4 +1,4 @@
-package modelo.alarma;
+package modelo_r.alarma;
 
 import javax.sound.sampled.Clip;
